frontend: move generic list-page hooks to lib/hooks (#831)
