+++ frontend/ui/src/app/projects/[projectId]/users/page.tsx
@@ -8,7 +8,8 @@ import { Workflow, Users, Layers } from "lucide-react";
 import { SearchFilterBar } from "@/components/search-filter-bar";
 import { ListPagination } from "@/components/list-pagination";
 import { ProjectBreadcrumb } from "@/features/projects/components";
-import { useUsers, useListPageState } from "@/features/traces/hooks";
+import { useUsers } from "@/features/traces/hooks";
+import { useListPageState } from "@/lib/hooks/use-list-page-state";
 import { formatDate, formatCost, formatTokens, cn, buildUrlWithFilters } from "@/lib/utils";
 import type { UserListItem } from "@/lib/api/users";
 import type { UserQueryOptions } from "@/lib/api/users";
